Create HTML Fundamentals Shoes Component
This was one of the challenges within this section of the course. The aim was to recreate a page component provided by the tutor. The component was showing the price and other information for a pair of Chuck Taylors. This was a component that will be used later as part of a larger webpage for an e-commerce site.